pyproject.toml
[project]
name = "ai-research-assistant-for-obsidian"
version = "0.2.0"
description = "Automated RSS pipeline that turns articles, videos, and podcasts into Obsidian notes using Claude Code"
requires-python = ">=3.11"

dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.5",
]

[project.scripts]
ai-research-assistant = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
